requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.13.1
groq==0.4.2
langchain==0.1.0
langchain-community==0.0.10
sentence-transformers==2.2.2
transformers==4.36.0
faiss-cpu==1.13.2
numpy==1.26.2
torch==2.2.2
scikit-learn==1.3.2
pandas==2.1.3
tqdm==4.66.1
ijson==3.2.3
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0
